Add createVault rpc
